Replace internal colors of the ColorsGenerators and window's background color to ImVec4
The round trip sf::Color --> ImVec4 --> sf::Color doesn't work well with the ColorPicker : when selecting a black color the hue resetted itsef to red. Using only ImVec4 removes this problem. The ColorGenerator interface was not modified, only the internals. I feared that the conversion between ImVec4 and sf::Color would have some performance degradation for big L-Systems using a lot of LinearGradient, but measurements proves that there is no significant differences in one way or the other