Refactor the SaveMenu into its own class and add external popup management
The save menu was a monstrosity of a function with tons of static variables and redundant code. To clean it up a little, I put it into a separate class, which is still convoluted, but at least separated into several sub-function with static variables reduced and the remaining put into private members.

The second addition is the PopupGUI class and associatd functions: when realizing the LSystemView size warning popup should also respond to keyboard presses, add a separate popup management now seem necessary. The implementation is quite simple but with some edge cases like popup opening popup then closing, which should first close the popup than add the new one into the stack.
It has also subtle behaviors like the Observer framework, and the SaveMenu behavior was a little affected.