HUGE COMMIT: big refactoring of RuleMapBuffer
Even after the last modification, I just couldn't code on this class withouth pulling my hair out. So, big refactoring: RuleMapBuffer stopped being a Observer of a RuleMap, but has a RuleMap as a member through composition. As such, the old model 1 x RuleMap->n x RuleMapBuffer is over, as that's why all the complications came from, as a RuleMapBuffer updated the RuleMap which notified the LSystemView and all the others RuleMapBuffer... So the base RuleMapBuffer was being modified at the same time as it was modifying the RuleMap, and as such pointer invalidations was a huge problem.
Now, the RuleMapBuffers are Observable, so LSystemView stopped observing the two RuleMaps and having RuleMapBuffers at the same time, but simply observes the revised RuleMapBuffers. The RuleMaps now are never observed, and that's okay.

I added a lot of tests during all these steps, and RuleMapBuffer was a problem, as the infernal callback modification somehow fixed some bugs that were in the code. Now that it's gone, the bugs came back, and so there is fixes and new tests.

As a bonus, the 'revert()' and 'validate()' mecanism works as intended.